[master] merge libiscpk11 to libisc
3735.	[cleanup]	Merged the libiscpk11 library into libisc
			to simplify dependencies. [RT #35205]

+++ lib/isc/win32/time.c
@@ -333,7 +333,7 @@ isc_time_parsehttptimestamp(char *buf, isc_time_t *t) {
 	when = isc_tm_timegm(&t_tm);
 	if (when == -1)
 		return (ISC_R_UNEXPECTED);
-	isc_time_set(t, when, 0);
+	isc_time_set(t, (unsigned int)when, 0);
 	return (ISC_R_SUCCESS);
 }
